+++ lib/medusa/core.rb
@@ -202,7 +202,7 @@ module Medusa
       storage = Storage::Base.new(@opts[:storage] || Storage.Moneta(:Memory))
       storage.clear if @opts[:clear_on_startup]
       @pages = PageStore.new(storage)
-      @opts[:cookie_store] = CookieStore.new(@opts[:cookies], origins: @urls)
+      @opts[:cookie_store] = CookieStore.new(@opts[:cookies])
       @robots = Robotex.new(@opts[:user_agent]) if @opts[:obey_robots_txt]
       cache = HTTP::Cache.from(@opts[:http_cache], logger: @opts[:logger])
       @opts[:http_cache] = cache if cache
@@ -217,7 +217,7 @@ module Medusa
     #
     def freeze_options
       @opts.except(:logger, :http_cache, :cookie_store).each_value(&:freeze)
-      @opts[:cookies]&.each_key { @opts[:cookies][_1].freeze }
+      @opts[:cookies]&.each_value(&:freeze) if @opts[:cookies].is_a?(Hash)
       @opts.freeze
     end